requirements.txt
discord.py
pyyaml
emoji
pandas
matplotlib
